==> 1.Harvesting&Cleaning.Rmd <==
---
title: "ReplicationRMD"
author: "Alejandro Aísa"
date: "`r Sys.Date()`"
output: html_document
---

This notebook contains all the sub-steps of the first stage of the analysis in a single file. The main purpose is to unify the process so the replication code for the master's thesis is easier and more understandable. Small text will be added to each chunk of code to explain the logic of the step and/or clarify some issues.

## 1. Treating the raw data.

The first task done in the process of writing the thesis corresponded to the extraction and management of the data needed for the analysis: the data of protests and the data for the football matches.

```{r message=FALSE, warning=FALSE}
library(readr)
library(tidyverse)
```

### 1.1. Cleaning protests.

In order to get the data, I downloaded the files in [Ron Francisco's webpage](https://ronfran.ku.edu/data/index.html) concerning both Northern Ireland and Republic of Ireland. At the beginning of the process, the Republic of Ireland was to be included in the analysis. However, It was ultimately removed, but its data had already been cleaned. At this step, we will only prepare the data for the analysis; make the column names more understandable and add time and location.

#### Northern Ireland

First step is to include all the downloaded data in *~/Databases/Protests/Original*. From them, we may unify all the files into single dataframes referring to a region and store them in *~/Databases/Protests/* with the name `NI_Raw`.

```{r message=FALSE, warning=FALSE}
NI80_83 <- read_delim("Databases/Protests/Original/NI80_83.csv", 
                      delim = ";", escape_double = FALSE, trim_ws = TRUE)
NI84_86 <- read_delim("Databases/Protests/Original/NI84_86.csv", 
                      delim = ";", escape_double = FALSE, trim_ws = TRUE)
NI87_89 <- read_delim("Databases/Protests/Original/NI87_89.csv", 
                      delim = ";", escape_double = FALSE, trim_ws = TRUE)
NI90_92 <- read_delim("Databases/Protests/Original/NI90_92.csv", 
                      delim = ";", escape_double = FALSE, trim_ws = TRUE)
NI93_95 <- read_delim("Databases/Protests/Original/NI93_95.csv", 
                      delim = ";", escape_double = FALSE, trim_ws = TRUE)

NI_Raw <- rbind(NI80_83, NI84_86, NI87_89,
              NI90_92, NI93_95)

write.csv(NI_Raw, file = "Databases/Protests/NI_Raw.csv")
```

Once the data is merged, we can rename some date columns, create others, and add the district for the different population, all with the aim of standardization.  

As a small warning, the following part was originally done in a different script. Moreover, it is possible that some locations are wrongly distributed. This step was done by hand, using the 26 districts that used to divide the region between 1972 and 2015 as reference. Also, the issues non-related to *The Troubles* are defined at this stage. The same issue may appear.  

```{r}
Antrim <- c("Antrim", "Toomebridge", "Parkgate", "Red Arch", "Derryiaghy") 
Ards <- c("Newtownards")
Armagh <- c("Armaugh", "Armagh", "South Armaugh", "Newtownhamilton", 
            "Millford", "Keady", "Markethill", "Darkley", "Darkley",
            "Portadown", "Tynan", "Loughgall", "Benburb", "Carrickarnan", 
            "Middletown", "Craigavon", "Jonesborough", "Glenanne", "County Armagh",
            "Charlemont", "Bull's Cross", "Ballyargan")
Ballymena  <- c("Ballymena")
Ballymoney  <- c("Rasharkin", "Dunloy", "Ballymoney")
Banbridge  <- c("Banbridge")
Belfast  <- c("Belfast", "West Belfast", "New Barnsley", "Ballymurphy",
              "Maryfield", "Anderstown","Rathcoole", "Andersonstown", 
              "Queen's Island", "Dunmurry", "Crumlin", "Ardoyne", "Newtownbreda", 
              "Andersontown", "Ballynafeigh" )
Carrickfergus  <- c("Carrickfergus")
Castlereagh  <- c("Castlereagh")
Coleraine  <- c("Portrush", "Coleraine", "Victoria Bridge", "Kilrea")
Cookstown  <- c("Cookstown", "Tullylagan", "Pomeroy", "Ardboe", "Stewartstown") 
Craigavon  <- c("Lurgan", "Trandragee")
Derry  <- c("Derry", "Lough Foyle", "Magherafelt", "Limavady", "Derryville", 
            "Derrylin", "Lecky Bridge", "Bellaghy", "Creggan", "Ballinderry", 
            "Craggan")
Down  <- c("Castlewellan", "Warrenpoint Harbor",
           "Rostrevor", "Warrenpoint", "Casstlewellan", "Ballynahinch", 
           "Downpatrick", "South Down", "Annalong", "Annalong", 
           "Newcastle", "County Down", "Dundrum", "Crossgar", "Hollywood", 
           "Annacloy")
Dungannon <- c("Derrylauchan", "Dungannon", "Ballygawley", "Coal Island", 
               "Pomeroy", "Coalisland", "Aughnacloy", "Eglish", 
               "Moy", "Augher", "Clogher", "Clonoe", "Caledon", "Fivemiletown", 
               "Dernanaught")
Fermanagh <- c("Fermanagh", "Magheraveely", "Enniskillen", "County Fermanagh", 
               "Maguiresbridge", "Beleek", "Rosslea", "Tullyhommon", 
               "Lisnaskea", "Loughmacory", "Six Mile Cross", "Kinawley", 
               "Brookeborough", "Belleek", "Mullybritt", "Linsaskea",  
               "Annaghmartin", "Derryvollen", "Monaghan", "Cloughoge", 
               "Lackey bridge","Clonnaty Bridge", "Kilturk", "Roslea")
Larne <- c("Larne")
Limavady <- c("Limavady", "Garvagh")
Lisburn <- c("Lisburn", "Moira", "Hillsborough", "Long Kesh")
Magherafelt <- c("Draperstown", "Moneymoor", "Ballyronan", "Coagh","Maghera", 
                 "Swatragh", "Moneymore")
Moyle <- c("Ballycastle")
`Newry & Mourne` <- c("Newry", "Forkill", "Crossmaglen", "Comlough", 
                 "Beesbrook", "Kilnasagart", "Kilkeel", "Killeen", 
                 "Camlough", "Mayobridge", "Silverbridge","Cullyhanna", 
                 "Forkhill", "Cullaville", "Whitecross", "Newtownhamiltion")
Newtownabbey <- c("Newtownabbey", "Glengormley")
NorthDown <- c("Holywood", "Bangor")
Omagh <- c("Tyrone", "Carrickmore", "Omagh", "County Tyrone","Galbally",
           "Drumnakilly", "Drumquin", "Cappagh", "Fintona")
Strabane <- c("Strabane", "Clady", "Castlederg", "Sion Mills", 
              "Victoria Bridge", "Newtownstewart", "Killeter","Newtonstewart",
              "Ballymagorry", "Donemana")

Other <- c("Irish Border", "Suffolk", "Carcross", "Northern Ireland",
          "Scotstown", "Irish border", "Aughinlig", "Donaghamore", 
          "King's Island",
          "Belcool", "Collon", "Bessbrook", "Lough Neagh")


IssuesRemoval <- c("wages", "meal money", "ferry service", "farm prices", "labor agreement", 
                   "bonus payments", "unemployment compensation", "worker rights", 
                   "unemployment", "shareholder rights", "travel payments", 
                   "medical insurance", "pornography", "pay agreement", "worker benefits", 
                   "fishing rights", "pension rights", "tax protest", "working hours", 
                   "labor benefits", "nuclear disarmament", "rock & roll", "subsidies", 
                   "milk production", "travel expenses", "ground rent", "lax security", 
                   "U.S. foreign policy", "travel subsidies", "worker solidarity", 
                   "animal rights", "drug use", "changing work", "bonuses", "AIDS segregation", 
                   "flight delays", "anti-alcohol", "electrical interference", "ecology", 
                   "fishing tax", "anti-homosexual law", "privatization", "working conditions", 
                   "anti-supermarkets", "price war", "anti-nuclear weapons", "price freeze", 
                   "workers' cooperative", "teacher rights", "condom sales", "private buses",
                   "women clergy", "severance money", "farm subsidies", "farm imports",
                   "bankruptcy", "private competition", "farm payments", "gay rights", 
                   "women's rights", "university funds", "tax on clothing", "child abuse",
                   "U.S. bombing", "phone fees", "maternity ward", "East Timor rights", 
                   "crop subsidies", "cattle subsidies", "nuclear reprocessing",
                   "Christmas leave", "animal cruelty", "anti-nuclear", "fishing subsidies", 
                   "pork subsidies", "anti-abortion", "work grades", "course dropped", 
                   "school delay", "abortion rights", "ocean pollution", "bigger zoo", 
                   "pork prices", "nursing prices", "pensions", "U.S. death penalty", 
                   "student aid", "drunk driving law", "livestock export", "driver safety", 
                   "early retirement", "farmer rights", "nuclear tests", "EU cycle rules",
                   "radio tower", "sheep prices", "anti-privatization", "drug treatment center", 
                   "milk prices", "elderly care home", "anti-divorce", "pro-divorce", 
                   "genetic engineering", "restructuring", "pollution")
```

```{r}
NIProtestsDF <- NI_Raw %>% 
  mutate(
    Date = dmy(Date)) %>% 
  rename(
    Perpetrator = Protester, 
    Political_target = `State/target`,
    Target = Agent, 
    N_Protesters = `# Protesters`,
    N_ProtestersInjured = Injuries...15, 
    N_StateInjured = Injuries...19, 
    N_StateForces = `# State`) %>% 
  mutate( 
    District = case_when( 
      Location %in% Antrim ~ "Antrim", 
      Location %in% Ards ~ "Ards",
      Location %in% Armagh ~ "Armagh",
      Location %in% Ballymena ~ "Ballymena",
      Location %in% Ballymoney ~ "Ballymoney",
      Location %in% Banbridge ~ "Banbridge",
      Location %in% Belfast ~ "Belfast",
      Location %in% Carrickfergus ~ "Carrickfergus",
      Location %in% Castlereagh ~ "Castlereagh",
      Location %in% Coleraine ~ "Coleraine",
      Location %in% Cookstown ~ "Cookstown",
      Location %in% Craigavon ~ "Craigavon",
      Location %in% Down ~ "Down", 
      Location %in% Derry ~ "Derry",
      Location %in% Dungannon ~ "Dungannon",
      Location %in% Fermanagh ~ "Fermanagh",
      Location %in% Larne ~ "Larne",
      Location %in% Limavady ~ "Limavady",
      Location %in% Lisburn ~ "Lisburn",
      Location %in% Magherafelt ~ "Magherafelt",
      Location %in% Moyle ~ "Moyle",
      Location %in% Newtownabbey ~ "Newtownabbey",
      Location %in% NorthDown ~ "North Down",
      Location %in% Omagh ~ "Omagh",
      Location %in% `Newry & Mourne` ~ "Newry & Mourne",
      Location %in% Strabane ~ "Strabane",
      Location %in% Other ~ "Other"))

write.csv(NIProtestsDF, file = "Databases/Protests/NIProtestsDF.csv")
```

#### Republic of Ireland

Same steps as before:

```{r message=FALSE, warning=FALSE}
Ireland80_83 <- read_delim("Databases/Protests/Original/Ireland80_83.csv", 
                           delim = ";", escape_double = FALSE, trim_ws = TRUE)
Ireland84_86 <- read_delim("Databases/Protests/Original/Ireland84_86.csv", 
                           delim = ";", escape_double = FALSE, trim_ws = TRUE)
Ireland87_89 <- read_delim("Databases/Protests/Original/Ireland87_89.csv", 
                           delim = ";", escape_double = FALSE, trim_ws = TRUE)
Ireland90_92 <- read_delim("Databases/Protests/Original/Ireland90_92.csv", 
                           delim = ";", escape_double = FALSE, trim_ws = TRUE)
Ireland93_95 <- read_delim("Databases/Protests/Original/Ireland93_95.csv", 
                           delim = ";", escape_double = FALSE, trim_ws = TRUE)

Ireland_Raw <- rbind(Ireland80_83, Ireland84_86, Ireland87_89, 
                   Ireland90_92, Ireland93_95) 

write.csv(Ireland_Raw, file = "Databases/Protests/Ireland_Raw.csv")

IrelandProtestsDF <- Ireland_Raw %>% 
   mutate(
     Date = dmy(Date)) %>% 
  rename(
    Perpetrator = Protester, 
    Political_target = `State/target`,
    Target = Agent, 
    N_Protesters = `# Protesters`,
    N_ProtestersInjured = Injuries...15, 
    N_StateInjured = Injuries...19, 
    N_StateForces = `# State`) %>% 
  mutate(
    District = "TBD"
  )
   
write.csv(IrelandProtestsDF, file = "Databases/Protests/IrelandProtestsDF.csv")
```

#### Merging both territories and filtering

Once we have the data cleaned and more understandable, I merged both territories and filtered out the protests that were not related to the *The Troubles*. Also, I removed the prison issues as I contemplate that they cannot be influenced by the outcome of matches. 

```{r}
ProtestsDF <- rbind(IrelandProtestsDF, NIProtestsDF) 

write.csv(ProtestsDF, file = "Databases/Protests/ProtestsDF.csv")

ProtestsDF <- read_csv("Databases/Protests/ProtestsDF.csv") %>% select(-...1) 

ProtestsFiltered <- ProtestsDF %>% 
  arrange(Date) %>% 
  drop_na(Event) %>% 
  filter(Target != "prison" & Target != "prison guards") %>% 
  filter(!(Issue %in% IssuesRemoval)) 

write.csv(ProtestsFiltered, file = "Databases/Protests/ProtestsFiltered.csv")
```

### 1.2. Cleaning football data

#### 1.2.1. Derry City FC

```{r}
library(httr2)
library(scrapex)
library(rvest)
library(httr)
library(tidyverse)
library(jsonlite)
library(xml2)
```

The main problem with the matches of Derry City lies on the fact that I had not had access to them as a unified dataframe or entrance. Thus, I had to first extract them from [Extratime.com](https://www.extratime.com/teams/results/7/derry-city/?results=1987) year by year. For that, I created the following custom function:

```{r}
derry_extractor <- function(x){ 
  
  Sys.sleep(15)
  
  HTML <- x %>% read_html()

  competitionDF <- HTML %>% 
    xml_find_all("//td[.//a[contains(@href, '/competition')]]") %>% 
    xml_text() %>% 
    as_tibble() %>% 
    rename(Competition = value)
  
  dateDF <- HTML %>% 
    xml_find_all("//tr//td[2]") %>% 
    xml_text() %>% 
    str_extract(".+ [0-9][0-9]") %>% 
    as_tibble() %>% 
    drop_na() %>% 
    rename(Date = value)
  
  against <- HTML %>% 
    xml_find_all("//td[.//a[contains(@href, '/teams')]]") %>% 
    xml_text() %>% 
    str_replace_all("\n      \n        ", "") %>% 
    str_split("\n        ") 
  
  againstDF <- as_tibble(map_chr(against, 1)) %>% 
    rename(against = value)
  
  ground <- HTML %>% 
    xml_find_all("//tr//td[4]") %>% 
    xml_text() %>% 
    str_replace_all("\n        ", "") %>% 
    str_replace_all("\n       \n    ", "") %>% 
    str_replace_all("\n      ", "") %>% 
    as_tibble() %>% 
    rename(Ground = value)
  
  groundDF <- ground[-c(1:3), ]
  
  results <- HTML %>% 
    xml_find_all("//td[.//a[contains(@href, '/fixture')]]") %>% 
    xml_text() %>% 
    str_extract_all("[0-9] - [0-9]")
  
  ResultsDF <- data.frame(Results = do.call(rbind, results))
  
  DF <- cbind(competitionDF, dateDF, againstDF, groundDF, ResultsDF) 
      
}
```

This function allow me to extract the name of the competition, the date of the match, the opponent, the ground and the result. Instead of performing a loop over all the years of competition, I decided to apply the function individually for identification purposes.

```{r}
Derry1987 <- "https://www.extratime.com/teams/results/7/derry-city/?results=1987"
Derry1988 <- "https://www.extratime.com/teams/results/7/derry-city/?results=1988" 
Derry1989 <- "https://www.extratime.com/teams/results/7/derry-city/?results=1989" 
Derry1990 <- "https://www.extratime.com/teams/results/7/derry-city/?results=1990" 
Derry1991 <- "https://www.extratime.com/teams/results/7/derry-city/?results=1991" 
Derry1992 <- "https://www.extratime.com/teams/results/7/derry-city/?results=1992" 
Derry1993 <- "https://www.extratime.com/teams/results/7/derry-city/?results=1993" 
Derry1994 <- "https://www.extratime.com/teams/results/7/derry-city/?results=1994" 
Derry1995 <- "https://www.extratime.com/teams/results/7/derry-city/?results=1995"

seven <- derry_extractor(Derry1987)
eight <-  derry_extractor(Derry1988)
nine <- derry_extractor(Derry1989)
ten <- derry_extractor(Derry1990)
one <- derry_extractor(Derry1991)
two <- derry_extractor(Derry1992)
three <- derry_extractor(Derry1993)
four <- derry_extractor(Derry1994)
five <- derry_extractor(Derry1995)
```

Once we have the data for all the competition, I formatted the columns so I could later perform the cleaning operations. Then, I unified all the years into a single dataframe containing all the matches. The resulting dataframe is the equivalent to the protests one, before the cleaning (NI_Raw). 

```{r}
DF1987 <- seven %>% 
  separate(Date, c("Month", "Day"), sep = " ") %>% 
  mutate(
    Year = 1987, 
    Day = as.numeric(Day), 
    Month = as.numeric(
      case_when(
        Month == "Sep" ~ 09, 
        Month == "Oct" ~ 10,
        Month == "Nov" ~ 11,
        Month == "Dec" ~ 12,
        Month == "Jan" ~ 01, 
        Month == "Feb" ~ 02, 
        Month == "Mar" ~ 03, 
        Month == "Apr" ~ 04, 
        Month == "May" ~ 05, 
        Month == "Aug" ~ 08)), 
    Date = paste(Day, Month, Year, sep = "/"), 
    Date = dmy(Date))
      
DF1988 <- eight %>% 
  separate(Date, c("Month", "Day"), sep = " ") %>% 
  mutate(
    Year = 1988, 
    Day = as.numeric(Day), 
    Month = as.numeric(
      case_when(
        Month == "Sep" ~ 09, 
        Month == "Oct" ~ 10,
        Month == "Nov" ~ 11,
        Month == "Dec" ~ 12,
        Month == "Jan" ~ 01, 
        Month == "Feb" ~ 02, 
        Month == "Mar" ~ 03,
        Month == "Apr" ~ 04, 
        Month == "May" ~ 05, 
        Month == "Aug" ~ 08)), 
    Date = paste(Day, Month, Year, sep = "/"), 
    Date = dmy(Date))

DF1989 <- nine %>% 
  separate(Date, c("Month", "Day"), sep = " ") %>% 
  mutate(
    Year = 1989, 
    Day = as.numeric(Day), 
    Month = as.numeric(
      case_when(
        Month == "Sep" ~ 09, 
        Month == "Oct" ~ 10,
        Month == "Nov" ~ 11,
        Month == "Dec" ~ 12,
        Month == "Jan" ~ 01, 
        Month == "Feb" ~ 02, 
        Month == "Mar" ~ 03,
        Month == "Apr" ~ 04,
        Month == "May" ~ 05,  
        Month == "Aug" ~ 08)), 
    Date = paste(Day, Month, Year, sep = "/"), 
    Date = dmy(Date))

DF1989 <- seven %>% 
  separate(Date, c("Month", "Day"), sep = " ") %>% 
  mutate(
    Year = 1989, 
    Day = as.numeric(Day), 
    Month = as.numeric(
      case_when(
        Month == "Sep" ~ 09, 
        Month == "Oct" ~ 10,
        Month == "Nov" ~ 11,
        Month == "Dec" ~ 12,
        Month == "Jan" ~ 01, 
        Month == "Feb" ~ 02, 
        Month == "Mar" ~ 03,
        Month == "Apr" ~ 04,
        Month == "Aug" ~ 08)), 
    Date = paste(Day, Month, Year, sep = "/"), 
    Date = dmy(Date))

DF1990 <- ten %>% 
  separate(Date, c("Month", "Day"), sep = " ") %>% 
  mutate(
    Year = 1990, 
    Day = as.numeric(Day), 
    Month = as.numeric(
      case_when(
        Month == "Sep" ~ 09, 
        Month == "Oct" ~ 10,
        Month == "Nov" ~ 11,
        Month == "Dec" ~ 12,
        Month == "Jan" ~ 01, 
        Month == "Feb" ~ 02, 
        Month == "Mar" ~ 03,
        Month == "Apr" ~ 04,
        Month == "Aug" ~ 08)), 
    Date = paste(Day, Month, Year, sep = "/"), 
    Date = dmy(Date))

DF1991 <- one %>% 
  separate(Date, c("Month", "Day"), sep = " ") %>% 
  mutate(
    Year = 1991, 
    Day = as.numeric(Day), 
    Month = as.numeric(
      case_when(
        Month == "Sep" ~ 09, 
        Month == "Oct" ~ 10,
        Month == "Nov" ~ 11,
        Month == "Dec" ~ 12,
        Month == "Jan" ~ 01, 
        Month == "Feb" ~ 02, 
        Month == "Mar" ~ 03, 
        Month == "Apr" ~ 04,
        Month == "Aug" ~ 08)), 
    Date = paste(Day, Month, Year, sep = "/"), 
    Date = dmy(Date))

DF1992 <- two %>% 
  separate(Date, c("Month", "Day"), sep = " ") %>% 
  mutate(
    Year = 1992, 
    Day = as.numeric(Day), 
    Month = as.numeric(
      case_when(
        Month == "Sep" ~ 09, 
        Month == "Oct" ~ 10,
        Month == "Nov" ~ 11,
        Month == "Dec" ~ 12,
        Month == "Jan" ~ 01, 
        Month == "Feb" ~ 02, 
        Month == "Mar" ~ 03, 
        Month == "Apr" ~ 04,
        Month == "Aug" ~ 08)), 
    Date = paste(Day, Month, Year, sep = "/"), 
    Date = dmy(Date))

DF1993 <- three %>% 
  separate(Date, c("Month", "Day"), sep = " ") %>% 
  mutate(
    Year = 1993, 
    Day = as.numeric(Day), 
    Month = as.numeric(
      case_when(
        Month == "Sep" ~ 09, 
        Month == "Oct" ~ 10,
        Month == "Nov" ~ 11,
        Month == "Dec" ~ 12,
        Month == "Jan" ~ 01, 
        Month == "Feb" ~ 02, 
        Month == "Mar" ~ 03, 
        Month == "Apr" ~ 04,
        Month == "Aug" ~ 08)), 
    Date = paste(Day, Month, Year, sep = "/"), 
    Date = dmy(Date))

DF1994 <- four %>% 
  separate(Date, c("Month", "Day"), sep = " ") %>% 
  mutate(
    Year = 1994, 
    Day = as.numeric(Day), 
    Month = as.numeric(
      case_when(
        Month == "Sep" ~ 09, 
        Month == "Oct" ~ 10,
        Month == "Nov" ~ 11,
        Month == "Dec" ~ 12,
        Month == "Jan" ~ 01, 
        Month == "Feb" ~ 02, 
        Month == "Mar" ~ 03, 
        Month == "Apr" ~ 04,
        Month == "Aug" ~ 08)), 
    Date = paste(Day, Month, Year, sep = "/"), 
    Date = dmy(Date))

DF1995 <- five %>% 
  separate(Date, c("Month", "Day"), sep = " ") %>% 
  mutate(
    Year = 1995, 
    Day = as.numeric(Day), 
    Month = as.numeric(
      case_when(
        Month == "Sep" ~ 09, 
        Month == "Oct" ~ 10,
        Month == "Nov" ~ 11,
        Month == "Dec" ~ 12,
        Month == "Jan" ~ 01, 
        Month == "Feb" ~ 02, 
        Month == "Mar" ~ 03, 
        Month == "Apr" ~ 04,
        Month == "Aug" ~ 08)), 
    Date = paste(Day, Month, Year, sep = "/"), 
    Date = dmy(Date))

DerryCityDF <- rbind(DF1987, DF1988, DF1989, DF1990, 
                 DF1991, DF1992, DF1993, DF1994, DF1995)

write.csv(DerryCityDF, file = "Databases/DerryCity/DerryCityDF.csv")
```

Then, we can perform the same tasks as with the protests. We rename some columns, and added the time and location. With respect to the location, a binary variable is created referring to the match played in Derry. Also, the dummy accounting for a Derry win is created. These two would be the independent variables.

```{r}
DerryCityDF <- read_csv("Databases/DerryCity/DerryCityDF.csv") 
  
DerryCityCleaned <- DerryCityDF %>% 
  dplyr::select(-...1) %>% 
  mutate(
    HomeTeam = if_else(Ground == "The Ryan McBride Brandywell Stadium", 1, 0)) %>%  
  separate(Results, c("HomeScore","AwayScore"), sep = " - ") %>% 
  arrange(Date) %>% 
  transmute(
    Date, 
    Competition, 
    Rival = against, 
    HomeScore, 
    AwayScore, 
    Ground, 
    DerryHome = HomeTeam, 
    DerryWin = 
      case_when(
        (DerryHome == "1" & HomeScore > AwayScore) ~ 1,
        (DerryHome == "1" & HomeScore <= AwayScore) ~ 0, 
        (DerryHome == "0" & HomeScore >= AwayScore) ~ 0, 
        (DerryHome == "0" & HomeScore < AwayScore) ~ 1)) %>% 
  filter(Date > "1987/08/01") %>% 
  mutate(
    District = ifelse(
      Ground == "The Ryan McBride Brandywell Stadium", "Derry", "Other"
    )) 

write.csv(DerryCityCleaned, file = "Databases/DerryCity/DerryCityCleaned.csv") 
```

#### 1.2.2. Cliftonville FAC

As mentioned in the thesis, the footballing results were provided by Haydn Milligan. Thus, the original data cannot be posted in this repository. However, the full list can be visualized [here](https://www.irishleaguearchive.org/by-team). Assuming that the data is present, the cleaning process would be the same: renaming of some variables, adding time and location and the dummy independent variables:

```{r}
CliftonvilleCleaned <- Cliftonville80_95 %>% 
  mutate(MatchDate = matchDate) %>% 
  separate(matchDate, c("Year", "Month", "Day"), sep = "-") %>% 
  transmute(
    Season = season, 
    MatchDate = ymd(MatchDate), 
    Competition = competitionName, 
    Stage = stage, 
    HomeTeam = homeTeamName,
    HomeScore = homeScore,
    AwayTeam = awayTeamName,
    AwayScore = awayScore, 
    Ground = groundName, 
    CliftonvilleHome = if_else((Ground == "Solitude" | Ground == "-"), "1", "0"), 
    CliftonvilleWin = case_when(
        (CliftonvilleHome == "1" & HomeScore > AwayScore) ~ 1,
        (CliftonvilleHome == "1" & HomeScore <= AwayScore) ~ 0, 
        (CliftonvilleHome == "0" & HomeScore >= AwayScore) ~ 0, 
        (CliftonvilleHome == "0" & HomeScore < AwayScore) ~ 1)) %>% 
  filter(MatchDate > "1980-01-01") %>% 
  mutate(
    MatchID = 1:n(), 
    MatchID = paste0(MatchID, "C"), 
    District = case_when(
      Ground == "Solitude" ~ "Belfast",
      Ground == "Castlereagh Park" ~ "Ards", 
      Ground == "Windsor Park" ~ "Belfast",
      Ground == "Seaview" ~ "Belfast",
      Ground == "Ballymena Showgrounds" ~ "Ballymena",
      Ground == "Shamrock Park" ~ "Armagh",
      Ground == "Inver Park" ~ "Larne",
      Ground == "Mourneview Park" ~ "Craigavon",
      Ground == "St Mels Park" ~ "Other",
      Ground == "Coleraine Showgrounds" ~ "Coleraine",
      Ground == "Clandeboye Park" ~ "North Down",
      Ground == "The Oval" ~ "Belfast",
      Ground == "New Grosvenor Stadium" ~ "Lisburn",
      Ground == "Newry Showgrounds" ~ "Newry & Mourne",
      Ground == "Dixon Park" ~ "Antrim",
      Ground == "Taylor's Avenue" ~ "Carrickfergus",
      Ground == "Dennis Harvey Park" ~ "Larne",
      Ground == "Limavady Showgrounds" ~ "Limavady",
      Ground == "Fortwilliam Park" ~ "Magherafelt",
      Ground == "St Julian's Road" ~ "Omagh",
      Ground == "Robert Adams Park" ~ "Down",
      Ground == "Ferney Park" ~ "Fermanagh",
      Ground == "Crewe Park" ~ "Antrim",
      Ground == "-" ~ "Belfast"))

write.csv(CliftonvilleCleaned, file = "~/CliftonvilleCleaned.csv")
```

### 1.3. Cleaning control variables

#### 1.3.1 Socio-Economic

In order to get the data for the control variables, I resorted to the UK censuses of 1981 and 1991 and the Cain archives. In the first case, one can easily download the data from [nomis](https://www.nomisweb.co.uk/). Then, we just renamed the variables and aggregate some of them to obtain the percentage of households with car and the percentage of unemployed people.

```{r}
census1991 <- read_delim("Databases/1991census.csv", 
                          delim = ";", escape_double = FALSE, trim_ws = TRUE)
control1991 <- census1991 %>% 
  transmute( # Renaming 
    District =  Area,
    PopulationM = `S67:295 (Aged 16 and over : Total male )`, 
    PopulationF = `S67:296 (Aged 16 and over : Total female )`, 
    Population = `S02:1 (All ages : Total persons )`, 
    MalesActives = `S08:12 (Males Economically Active : Total Aged 16 and Over )`, 
    MalesUnem = `S08:78 (Males EA unemployed : Total Aged 16 and Over )`, 
    FemalesActives = `S08:166 (Females Economically Active : Total Aged 16 and Over )`, 
    FemalesUnem = `S08:232 (Females EA unemployed : Total Aged 16 and Over )`,
    TotalHH = `S21:2 (All households : Total households )`, 
    OneCarsHH = `S21:4 (All households : Households with 1 car )`, 
    TwoCarsHH = `S21:5 (All households : Households with 2 cars )`, 
    ThreeCarsHH = `S21:6 (All households : Households with 3 or more cars )`) %>% 
  mutate( 
    TotalActives = MalesActives + FemalesActives, 
    TotalUnem = MalesUnem + FemalesUnem, 
    pcg_Unem = TotalUnem / TotalActives, 
    pcg_car = (OneCarsHH + TwoCarsHH + ThreeCarsHH) / TotalHH)

write.csv(control1991, "Databases/control1991.csv")

census1981 <- read_delim("Databases/1981census.csv", 
                         delim = ";", escape_double = FALSE, trim_ws = TRUE)
  
control1981 <- census1981 %>% 
  transmute(
    District, 
    MalesActive, 
    FemalesActive,
    Population, 
    TotalActives = MalesActive + FemalesActive, 
    MalesUnem, 
    FemalesUnem, 
    TotalUnem = MalesUnem + FemalesUnem, 
    pcg_Unem = TotalUnem / TotalActives, 
    nocarHH, 
    carHH,
    TotalHH = nocarHH + carHH, 
    pcg_car = carHH/TotalHH)

write.csv(control1981, "Databases/control1981.csv")
```

With respect to the religious data, I collected it from [Cain](https://cain.ulster.ac.uk/ni/popul.htm#cath). Then, as usual, we rename the variables and create a dummy for those districts that possessed majority of catholics. While it would not be later used, it may had served in the analysis.

```{r}
ReligionLink <- "https://cain.ulster.ac.uk/ni/popul.htm#cath" %>% 
  read_html()

ReligionTables <- ReligionLink %>% html_table()

Pcg_catholics <- ReligionTables[[11]] 

names(Pcg_catholics) <- as.character(Pcg_catholics[1, ]) 

Pcg_catholics <- Pcg_catholics[-1, ]
Pcg_catholics <- Pcg_catholics[-27, ]

Religion <- Pcg_catholics %>% 
  mutate(
    District = `District Council Area`, 
    pcg1971 = as.numeric(`% Catholics 1971[1]`), 
    pcg1981 = as.numeric(`% Catholics 1981[2]`),
    pcg1991 = as.numeric(`% Catholics 1991[3]`),
    pcg2001 = as.numeric(`% Catholics 2001[4]`),
    pcg2011 = as.numeric(`% Catholics 2011[5]`)) %>% 
  select(District, pcg1981, pcg1991) %>% 
  mutate(
    CatholicMajority1981 = ifelse(pcg1981 > 50, 1, 0), 
    CatholicMajority1991 = ifelse(pcg1991 > 50, 1, 0))

write.csv(Religion, file = "Databases/Religion.csv")
```

#### 1.3.2. GIS data.

```{r}
library(sp)
library(rgdal)
library(rgeos)
library(sf)
library(tidyverse)
```

Finally, I downloaded the GIS data for the 26 districts from [GADM](https://gadm.org/data.html) in Geographic Information System (GIS) shapefile format and stored them in *\~/Data*. As the data incorporates all the United Kingdom and various administrative levels, I needed to keep only the region of Northern Ireland at the 26-district level. Then, I stored the new shapefile that accounts only for NI.

```{r}
UK <- readOGR(dsn = "data", layer = "gadm36_GBR_3") # until october 23

Districts <- c("Antrim", "Ards", "Armagh", "Ballymena", "Ballymoney", 
               "Banbridge", "Belfast", "Carrickfergus", "Castlereagh", 
               "Coleraine", "Cookstown", "Craigavon", "Derry", "Down", 
               "Dungannon", "Fermanagh", "Larne", "Limavady", "Lisburn", 
               "Moyle", "Newry and Mourne", "Newtownabbey", "North Down", 
               "Omagh", "Strabane", "Magherafelt")

NI <- UK[UK$NAME_3 %in% Districts, ]
```

Then, we can calculate the distances to Derry and Belfast from all of the other districts. It has to be noted here that each district is composed by a polygon, and the distances are calculated from the centroids of each one.

Thus, we may get the coordinates and centroids:

```{r message=FALSE, warning=FALSE}
NI_sf <- st_read(dsn = "data", layer = "NI")

district_coordinates <- NI_sf %>% 
  select(NAME_3, geometry)

district_center <- st_centroid(NI_sf) %>% 
  select(NAME_3, geometry) %>% 
  rename(
    center = geometry
  )

DCoor <- as.data.frame(district_coordinates)  
Dcentre <- as.data.frame(district_center)

```

Calculating the distances to Belfast:

```{r}
Belfast <- Dcentre %>% 
  filter(NAME_3 == "Belfast")

Belfast_distances <- DCoor %>% 
  left_join(Dcentre, by = "NAME_3") %>% 
  mutate(
    Belfast_point = Belfast$center, 
    Distance = st_distance(center, Belfast_point, by_element = T)) %>% 
  transmute(
    District = NAME_3, 
    Distance = as.numeric(Distance)) %>% 
  mutate(
    District = recode(District, "Newry and Mourne" = "Newry & Mourne"), 
    Distance_km = round((Distance / 1000), 2))

write.csv(Belfast_distances, "Databases/Belfast_Distances.csv")   
```

Calculating the distances to Derry:

```{r}
Derry <- Dcentre %>% 
  filter(NAME_3 == "Derry")

Derry_distances <- DCoor %>% 
  left_join(Dcentre, by = "NAME_3") %>% 
  mutate(
    Derry_point = Derry$center, 
    Distance = st_distance(center, Derry_point, by_element = T)) %>% 
  transmute(
    District = NAME_3, 
    Distance = as.numeric(Distance)) %>% 
  mutate(
    District = recode(District, "Newry and Mourne" = "Newry & Mourne"), 
    Distance_km = round((Distance / 1000), 2))

write.csv(Derry_distances, "Databases/Derry_Distances.csv")   
```

Now we have all the data ready to be merged and aggregated into a single dataframe. This step is covered in the markdown labelled as '2. Aggregating.rmd'.
