==> 4.1. AnimatedMaps.Rmd <==
---
title: "Animated Maps"
author: "Alejandro Aísa"
date: "`r Sys.Date()`"
output:
  tufte::tufte_html: default
  html_document:
    df_print: paged
---

```{r message=FALSE, warning=FALSE, include=FALSE}
library(sp)
library(rgdal)
library(rgeos)
library(sf)
library(lubridate)
require(gganimate)
require(transformr)
require(ggthemes)
library(ggspatial)
library(tidyverse)
require(tufte)
require(prettydoc)

```

```{r message=FALSE, warning=FALSE, include=FALSE}
NI <- read_sf("data/NI.shp")%>% 
  dplyr::select(NAME_3, geometry) %>% 
  rename(District = NAME_3) %>% 
  mutate(
    District = dplyr::recode(District, "Newry and Mourne" = "Newry & Mourne"))

df <- read_csv("Databases/Final_Derry.csv")

data_year <- df |> 
  dplyr::select(-...1) %>% 
  group_by(year = floor_date(week, unit = "year"), District) %>% 
  summarise(
    yearly_protests = sum(n_protests)) |> 
  ungroup()|> 
  mutate(
    year = year(year))

pyear <- NI |> 
  left_join(data_year, by = "District")

sysfonts::font_add_google("Gideon Roman", family = "gideon roman")
showtext::showtext_auto()
  

```

## Animated map 1: Total protests

The following map depicts the number of protests (log scale) in each year of the period of study. 

```{r message=FALSE, warning=FALSE, include=FALSE}
ma <- ggplot()+
  geom_sf(data = pyear, aes(fill = log(yearly_protests +1 )))+
  geom_sf_label(data = pyear, aes(label = District), size = 2.5)+
  scale_fill_gradient(low = "lightyellow", high = "red")+ 
  annotation_north_arrow(location = "tr", 
                         which_north = "true",
                         style = north_arrow_fancy_orienteering, 
                         height = unit(2, "cm"), 
                         width = unit(2, "cm"))+
  annotation_scale(location = "bl") +
  theme_map()+
  transition_manual(year)+
  ggtitle("Year: {frame + 1978}", 
          subtitle = "Number of protests per year")+
  labs(caption = "Source: Own work and Ron Francisco's database")+
  theme(
    plot.title = element_text(face = "bold", 
                              size = 14, 
                              hjust = 0.5, 
                              family = "gideon roman"), 
    plot.subtitle = element_text(size = 12, 
                                 hjust = 0.5, 
                                 family = "gideon roman"),
    plot.caption = element_text(size = 11, family = "gideon roman"), 
  legend.position = "none") +
  labs(x = NULL, y = NULL)
```

```{r echo=FALSE, message=FALSE, warning=FALSE, fig.align='center'}
pma <- animate(ma, fps = 5)
pma
```

## Animated map 2: Protests per capita 

The second version of the animated representations shows the evolution in time of the number of protests per 1000 people in each year. 

```{r message=FALSE, warning=FALSE, include=FALSE}
data_year_pc <- df |> 
  dplyr::select(-...1) %>% 
  group_by(year = floor_date(week, unit = "year"), District) %>% 
  reframe(
    protests_pc = sum(n_protests)/(Population/1000)) |> 
  unique() |> 
  mutate(
    year = year(year))

pyear_pc <- NI |> 
  left_join(data_year_pc, by = "District")
```

```{r message=FALSE, warning=FALSE, include=FALSE}
ma_pc <- ggplot()+
  geom_sf(data = pyear_pc, aes(fill = protests_pc))+
  geom_sf_label(data = pyear_pc, aes(label = District), size = 2.5)+
  scale_fill_gradient(low = "lightyellow", high = "red")+ 
  annotation_north_arrow(location = "tr", 
                         which_north = "true",
                         style = north_arrow_fancy_orienteering, 
                         height = unit(2, "cm"), 
                         width = unit(2, "cm"))+
  annotation_scale(location = "bl") +
  theme_map()+
  transition_manual(year)+
  ggtitle("Year: {frame + 1978}", 
          subtitle = "Number of protests per capita")+
  labs(caption = "Source: Own work and Ron Francisco's database")+
  theme(
    plot.title = element_text(face = "bold", 
                              size = 14, 
                              hjust = 0.5, 
                              family = "gideon roman"), 
    plot.subtitle = element_text(size = 12, 
                                 hjust = 0.5, 
                                 family = "gideon roman"),
    plot.caption = element_text(size = 10, family = "gideon roman"),
  legend.position = "none") +
  labs(x = NULL, y = NULL)
```

```{r echo=FALSE, message=FALSE, warning=FALSE, fig.align='center'}
pma_pc <- animate(ma_pc, fps = 5)
pma_pc
```


## Animated map 3: Protests per Catholic percentage

Finally, we can visualize the number of protests per year, dependant on the  percentage of catholics. 

```{r message=FALSE, warning=FALSE, include=FALSE}
data_year_pr <- df |> 
  dplyr::select(-...1) %>% 
  group_by(year = floor_date(week, unit = "year"), District) %>% 
  reframe(
    protests_pr = sum(n_protests)/(pcg_catholics/100)) |> 
  unique() |> 
  mutate(
    year = year(year))

pyear_pr <- NI |> 
  left_join(data_year_pr, by = "District")
```


```{r message=FALSE, warning=FALSE, include=FALSE}
ma_pr <- ggplot()+
  geom_sf(data = pyear_pr, aes(fill = protests_pr))+
  geom_sf_label(data = pyear_pr, aes(label = District), size = 2.5)+
  scale_fill_gradient(low = "lightyellow", high = "red")+ 
  annotation_north_arrow(location = "tr", 
                         which_north = "true",
                         style = north_arrow_fancy_orienteering, 
                         height = unit(2, "cm"), 
                         width = unit(2, "cm"))+
  annotation_scale(location = "bl") +
  theme_map()+
  transition_manual(year)+
  ggtitle("Year: {frame + 1978}", 
          subtitle = "Number of protests per catholic percentage")+
  labs(caption = "Source: Own work and Ron Francisco's database")+
  theme(
    plot.title = element_text(face = "bold", 
                              size = 14, 
                              hjust = 0.5, 
                              family = "gideon roman"), 
    plot.subtitle = element_text(size = 12, 
                                 hjust = 0.5, 
                                 family = "gideon roman"),
    plot.caption = element_text(size = 10, family = "gideon roman"),
    legend.position = "none") +
  labs(x = NULL, y = NULL)
```

```{r echo=FALSE, message=FALSE, warning=FALSE, fig.align='center'}
pma_pr <- animate(ma_pr, fps = 5)
pma_pr
```
