==> 5.Models.Rmd <==
---
title: "PoissonAnalysis"
author: "Alejandro Aísa"
date: "`r Sys.Date()`"
output: pdf_document 
---
```{r setup, include=FALSE}
knitr::opts_chunk$set(
	echo = TRUE,
	message = FALSE,
	warning = FALSE
)
```

The following code corresponds to the section II of the master's thesis. More specifically, it is devoted to the construction of various generalized linear models to analyse the effect of football matches in protests. As an clarifiying note, the interpretations of the results are present in the proper thesis. 

```{r message=FALSE, warning=FALSE}
library(caret)
library(pscl)
library(MASS)
library(msme)
library(car)
library(RcppRoll)
library(modelsummary)
library(kableExtra)
library(tidyverse)
```

## Models 

### Loading Data 

For this section of the thesis we load the database concerning the matches of Cliftonville. To allow the regression to work, some variables are recoded as factor. We also include the lagged variables for the negative binomial models. 

Similarly, matches are usually played on Sundays. Thus, we establish that the actual variable to predict is the number of protests of the following week. Also, some week-district had two matches. However, they cannot be removed as they entail different events, which can affect very differently to the proper outcome. 

```{r}
Cliftonville <- read.csv("Databases/Final_Cliftonville.csv") %>% 
  dplyr::select(-X) %>% 
  mutate(
    is_protest = as_factor(is_protest), 
    CliftonvilleHome = as_factor(CliftonvilleHome), 
    CliftonvilleWin = as_factor(CliftonvilleWin), 
    pcg_catholics = pcg_catholics/100) %>%  
  group_by(District) %>% 
  mutate(
    protests = lead(n_protests, n = 1), 
    lagged_month = roll_sum(n_protests, n = 3, 
                                         align = "right", fill = NA)) %>% 
  ungroup %>% 
  replace_na(
    list(
      protests = 0, 
      lagged_month = 0, 
      Distance_km = 46)) %>% 
  ungroup() %>% 
  dplyr::rename(
    lagged_week = n_protests)

```




## Description 

Before the analysis, it was necessary to get a glimpse of the data. 

### Binary 

```{r}
data.frame(table(Cliftonville$is_protest));histogram(Cliftonville$is_protest)
```

### Count 

```{r}
data.frame(table(Cliftonville$protests)); histogram(Cliftonville$protests);var(Cliftonville$protests)/mean(Cliftonville$protests) 

```

The previous information demonstrates the high number of week-districts without protests and the over dispersion of data. Thus, the need for Negative Binomial and Zero-Inflated models. 

### Keeping only numeric data 

We keep only numeric data to run the regressions: 

```{r}
data <- Cliftonville %>% 
  dplyr::select(-week, -District, -n_protesters) %>% 
  transmute(
    protests, 
    is_protest, 
    CHome = CliftonvilleHome, 
    CWin = CliftonvilleWin, 
    pcg_catholics, 
    pcg_car, 
    pcg_unem = pcg_Unem, 
    distance = Distance_km, 
    lagged_week,
    lagged_month)
  
```

## Models 

### Single model 

First model we train is a simple logistic regression accounting for the footballing variables. 

```{r}
log_sim <- glm(
  is_protest ~
    CHome + 
    CWin, 
  data = data, 
  family = "binomial")

summary(log_sim)

```

Secondly we can construct the same model, including the interaction between the variables. 

```{r}
log_int <- glm(
  is_protest ~ 
    CHome*CWin, 
  data = data, 
  family = "binomial")


summary(log_int)
```

### Logistic 

Lastly, we can train the logistic regression, including all the control variables: percentage of catholics, % oh Households with car, the unemployment level (%) and the distance to belfast. 

```{r}
log_com <- glm(
  is_protest ~
    CHome*CWin +
    pcg_catholics +
    pcg_car +
    pcg_unem + 
    distance,  
  data = data, 
  family = "binomial") 

summary(log_com); msme::P__disp(log_com)


```
Once the model is trained, we plot the increase in odds for protests for each term of the interaction. For that, we first create a new dataframe that accounts for the predicted odds, given average control variables. Then, we may calculate the limits for the confidence intervals. 

```{r}

nd = expand.grid(CHome = 0:1, CWin = 0:1) |> 
mutate(
  CHome = as.factor(CHome),
  CWin = as.factor(CWin))


nd$pcg_catholics = mean(data$pcg_catholics, na.rm = TRUE)
nd$pcg_car = mean(data$pcg_car, na.rm = TRUE)
nd$pcg_unem = mean(data$pcg_unem, na.rm = TRUE)
nd$distance = mean(data$distance, na.rm = TRUE)

nd$y = predict(log_com, newdata = nd)
nd$se = predict(log_com, newdata = nd, se.fit=T)$se.fit
nd$upr = nd$y + qnorm(0.975) * nd$se
nd$lwr = nd$y - qnorm(0.975) * nd$se
```


When we have the data, we just need to plot the odds: 

```{r}
ggplot(nd, aes(x = interaction(CHome, CWin), y = y)) +
  geom_pointrange(data = nd, aes(ymin = lwr, ymax = upr), color = "black", alpha = 0.3)+
  labs(
    title = "Figure 6: Predicted odds",
    subtitle = "Interaction term with complete logistic model",
    x = "Interaction", 
    y = "Predicted Odds", ) +
  scale_x_discrete(labels = c("Home=0, Win=0", "Home=1, Win=0", "Home=0, Win=1", "Home=1, Win=1"))+
  theme_classic() +
  theme(
    plot.title = element_text(face = "bold", 
                              size = 12, 
                              hjust = 0.5, 
                              family = "gideon roman"), 
    plot.subtitle = element_text(size = 10, 
                                 hjust = 0.5, 
                                 family = "gideon roman"), 
    axis.title.x = element_text(size = 10, family = "gideon roman"),
    axis.title.y = element_text(size = 10, family = "gideon roman"))
```



### Negative Binomial 

The negative binomial model is trained with all the previous variables + the lagged variables. Additionally, we use the *glm.nb* function form the `MASS` package. 

For mere statistical purposes, we may calculate the dispersion value of the model with the *P__disp* function from `msme` package. 

```{r}
NB <- glm.nb(
  protests ~
    CHome*CWin +
    pcg_catholics +
    pcg_car +
    pcg_unem + 
    distance + 
    lagged_week + 
    lagged_month, 
  data = data
)

summary(NB); P__disp(NB)

```


### Zero-Inflated Negative Binomial

Similarly, we can finally run the zero inflated negative binomial model with the *zeroinfl* function from `pscl`package. 

```{r}
data2 <- data %>% dplyr::select(-is_protest)

ZI_NB <- zeroinfl(
  protests ~
    CHome*CWin +
    pcg_catholics +
    pcg_car +
    pcg_unem +
    distance + 
    lagged_week + 
    lagged_month |
    CHome + 
    CWin +
    pcg_catholics +
    pcg_car +
    pcg_unem + 
    distance, 
  data = data2, 
  dist="negbin", 
  link="logit")

summary(ZI_NB); P__disp(ZI_NB)

```

### Additional models

Taking advantage of the data already loaded, we may run some other models. They are not included in the analysis, but they were initially considered. 

#### Quasi Poisson Model 

Due to the over-dispersion of data, a quasi-poisson model may be trained:

```{r}
QP_NI <- glm(
  protests ~
    CHome + 
    CWin +
    CHome*CWin +
    pcg_catholics +
    pcg_car +
    pcg_unem + 
    distance + 
    lagged_week + 
    lagged_month,  
  data = data, 
  family = "quasipoisson", 
  control = list(maxit = 100))

summary(QP_NI); P__disp(QP_NI) 
```

#### Zero-Inflated Poisson

The simpler version of the ZI-NB can be trained as follows: 

```{r}
ZI_NI2 <- zeroinfl(
  protests ~
    CHome*CWin + 
    pcg_catholics +
    pcg_car +
    pcg_unem +
    distance + 
    lagged_week + 
    lagged_month |
    CHome + 
    CWin +
    pcg_catholics +
    pcg_car +
    pcg_unem + 
    distance, 
  data = data, 
  dist="poisson", 
  link="logit")

summary(ZI_NI2); P__disp(ZI_NI2)
```
